// client/src/state/action-types/UsersActionTypes.ts
export enum UsersActionType {
  GETALL_STARTED = "GETALL_STARTED",
  GETALL_SUCCESS = "GETALL_SUCCESS",
  GETALL_FAILURE = "GETALL_FAILURE",

  SET_CURRENT_USER = "SET_CURRENT_USER",

  ADDUSER_STARTED = "ADDUSER_STARTED",
  ADDUSER_SUCCESS = "ADDUSER_SUCCESS",
  ADDUSER_FAILURE = "ADDUSER_FAILURE",

  UPDATEUSER_STARTED = "UPDATEUSER_STARTED",
  UPDATEUSER_SUCCESS = "UPDATEUSER_SUCCESS",
  UPDATEUSER_FAILURE = "UPDATEUSER_FAILURE",

  DELETEUSER_STARTED = "DELETEUSER_STARTED",
  DELETEUSER_SUCCESS = "DELETEUSER_SUCCESS",
  DELETEUSER_FAILURE = "DELETEUSER_FAILURE"

}